refactor: extract monitor routes and pages
