+++ frontend/monitor/src/test/setup.ts
@@ -1,1 +1,21 @@
 import "@testing-library/jest-dom/vitest";
+
+const storage = new Map<string, string>();
+
+Object.defineProperty(window, "localStorage", {
+  value: {
+    getItem(key: string) {
+      return storage.has(key) ? storage.get(key)! : null;
+    },
+    setItem(key: string, value: string) {
+      storage.set(key, value);
+    },
+    removeItem(key: string) {
+      storage.delete(key);
+    },
+    clear() {
+      storage.clear();
+    },
+  },
+  configurable: true,
+});
